% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatsAPIclientR-package.R
\docType{package}
\name{pcatsAPIclientR-package}
\alias{pcatsAPIclientR}
\alias{pcatsAPIclientR-package}
\title{pcatsAPIclientR: 'PCATS' API Client}
\description{
Provides an R interface to the 'PCATS' API \url{https://pcats.research.cchmc.org/api/__docs__/}, allowing R users to submit tasks and retrieve results.
}
\details{
The PCATS application programming interface (API) implements two Bayesian's non parametric causal inference modeling, Bayesian's Gaussian process regression and Bayesian additive regression tree, and provides estimates of averaged causal treatment (ATE) and conditional averaged causal treatment (CATE) for adaptive or non-adaptive treatment. The API is able to handle general types of treatment - binary, multilevel, continuous and their combinations, as well as general type of outcomes including bounded summary scores such as health related quality of life and survival outcomes. In addition, the API is able to deal with missing data using user supplied multiply imputed missing data. Summary tables and interactive figures of the results are generated and downloadable.
}
\keyword{internal}
