% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploturl.R
\name{ploturl}
\alias{ploturl}
\title{Return plot URL}
\usage{
ploturl(jobid, plottype = "")
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}

\item{plottype}{Plot type}
}
\value{
url
}
\description{
Return plot URL
}
