% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print job results}
\usage{
print(jobid)
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}
}
\value{
formatted text
}
\description{
Return formatted string with job results
}
