% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printgp.R
\name{printgp}
\alias{printgp}
\title{Print job results}
\usage{
printgp(jobid, token = NULL)
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}

\item{token}{Authentication token.}
}
\value{
formatted text
}
\description{
Return formatted string with job results
}
