% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Return job results}
\usage{
results(jobid, token = NULL)
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}

\item{token}{Authentication token.}
}
\value{
results
}
\description{
Return job results
}
