% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadfile.R
\name{uploadfile}
\alias{uploadfile}
\title{Upload a file}
\usage{
uploadfile(filename, token = NULL)
}
\arguments{
\item{filename}{Filename of a file to upload}

\item{token}{Authentication token.}
}
\value{
Backend filename reference
}
\description{
Upload a file
}
