% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_result.R
\name{wait_for_result}
\alias{wait_for_result}
\title{Wait while the job status is pending}
\usage{
wait_for_result(jobid)
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}
}
\value{
status
}
\description{
Return when the job status is finished (either successfully or otherwise)
}
