% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicGP.R
\name{dynamicGP}
\alias{dynamicGP}
\title{Fit a linear model with Gaussian Process at each stage for data with two time points.}
\usage{
dynamicGP(
  datafile = NULL,
  dataref = NULL,
  stg1.outcome,
  stg1.treatment,
  stg1.x.explanatory = NULL,
  stg1.x.confounding = NULL,
  stg1.tr.hte = NULL,
  stg1.tr.values = NULL,
  stg1.tr.type = "Discrete",
  stg1.outcome.type = "Continuous",
  stg1.outcome.bound_censor = "neither",
  stg1.outcome.lb = NULL,
  stg1.outcome.ub = NULL,
  stg1.outcome.censor.lv = NULL,
  stg1.outcome.censor.uv = NULL,
  stg1.outcome.censor.yn = NULL,
  stg1.outcome.link = "identity",
  stg2.outcome,
  stg2.treatment,
  stg2.x.explanatory = NULL,
  stg2.x.confounding = NULL,
  stg2.tr.hte = NULL,
  stg2.tr.values = NULL,
  stg2.tr.type = "Discrete",
  stg2.outcome.type = "Continuous",
  stg2.outcome.bound_censor = "neither",
  stg2.outcome.lb = NULL,
  stg2.outcome.ub = NULL,
  stg2.outcome.censor.lv = NULL,
  stg2.outcome.censor.uv = NULL,
  stg2.outcome.censor.yn = NULL,
  stg2.outcome.link = "identity",
  burn.num = 1000,
  mcmc.num = 1000,
  x.categorical = NULL,
  method = "BART",
  mi.datafile = NULL,
  mi.dataref = NULL,
  sheet = NULL,
  mi.sheet = NULL
)
}
\arguments{
\item{datafile}{File to upload (.csv or .xls)}

\item{dataref}{Reference to already uploaded file.}

\item{stg1.outcome}{The name of the outcome variable for stage 1.}

\item{stg1.treatment}{The name of the treatment variable for stage 1.}

\item{stg1.x.explanatory}{An optional vector of the names of the explanatory variables for stage 1.}

\item{stg1.x.confounding}{A vector of the name of the confounding variables for stage 1.}

\item{stg1.tr.hte}{An optional vector specifying categorical variables which may have heterogeneous treatment effect with the treatment variable for stage 1.}

\item{stg1.tr.values}{User-defined values for the calculation of ATE if the treatment variable is continuous for stage 1.}

\item{stg1.tr.type}{stg1.tr.type}

\item{stg1.outcome.type}{Outcome type ("Continuous" or "Discrete") for stage 1.}

\item{stg1.outcome.bound_censor}{stg1.outcome.bound_censor}

\item{stg1.outcome.lb}{Stage 1 lower bound if \code{stg1.outcome.censor} is TRUE.}

\item{stg1.outcome.ub}{Stage 2 upper bound if \code{stg2.outcome.censor} is TRUE.}

\item{stg1.outcome.censor.lv}{stg1.outcome.censor.lv}

\item{stg1.outcome.censor.uv}{stg1.outcome.censor.uv}

\item{stg1.outcome.censor.yn}{stg1.outcome.censor.yn}

\item{stg1.outcome.link}{stg1.outcome.link}

\item{stg2.outcome}{The name of the outcome variable for stage 2.}

\item{stg2.treatment}{The name of the treatment variable for stage 2.}

\item{stg2.x.explanatory}{An optional vector of the names of the explanatory variables for stage 2.}

\item{stg2.x.confounding}{A vector of the name of the confounding variables for stage 2.}

\item{stg2.tr.hte}{An optional vector specifying categorical variables which may have heterogeneous treatment effect with the treatment variable for stage 2.}

\item{stg2.tr.values}{User-defined values for the calculation of ATE if the treatment variable is continuous for stage 2.}

\item{stg2.tr.type}{stg2.tr.type}

\item{stg2.outcome.type}{Outcome type ("Continuous" or "Discrete") for stage 2.}

\item{stg2.outcome.bound_censor}{stg2.outcome.bound_censor}

\item{stg2.outcome.lb}{Stage 2 lower bound if \code{stg2.outcome.censor} is TRUE.}

\item{stg2.outcome.ub}{Stage 2 upper bound if \code{stg2.outcome.censor} is TRUE.}

\item{stg2.outcome.censor.lv}{stg2.outcome.censor.lv}

\item{stg2.outcome.censor.uv}{stg2.outcome.censor.uv}

\item{stg2.outcome.censor.yn}{stg2.outcome.censor.yn}

\item{stg2.outcome.link}{stg2.outcome.link}

\item{burn.num}{numeric; the number of MCMC 'burn-in' samples, i.e. number of MCMC to be discarded.}

\item{mcmc.num}{numeric; the number of MCMC samples after 'burn-in'.}

\item{x.categorical}{x.categorical}

\item{method}{method}

\item{mi.datafile}{File to upload (.csv or .xls) that contains the imputed data in the model.}

\item{mi.dataref}{Reference to already uploaded file that contains the imputed data in the model.}

\item{sheet}{If \code{dataurl} points to Excel file this variable specifies which sheet to load.}

\item{mi.sheet}{If \code{mi.dataurl} points to Excel file this variable specifies which sheet to load.}

\item{stg1.outcome.censor}{logical; if TRUE, stage 1 outcomes are bounded. If FALSE, outcomes are not bounded.}

\item{stg2.outcome.censor}{logical; if TRUE, stage 2 outcomes are bounded. If FALSE, outcomes are not bounded.}

\item{categorical}{A vector of the name of the categorical variables.}
}
\value{
jobid
}
\description{
FIXME: GPMatch is used to fit a linear model within a Bayesian framework.
  Gaussian process (GP) prior covariance function is utilized as a matching tool
  which accomplishes matching and flexible outcome modeling in a single step.

Return the GP dynamic results.
}
