% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status.R
\name{job_status}
\alias{job_status}
\title{Return job status}
\usage{
job_status(jobid, token = NULL)
}
\arguments{
\item{jobid}{Job ID of the previously submitted job}

\item{token}{Authentication token.}
}
\value{
status
}
\description{
Return status of the previously submitted job
}
