% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticGP.R
\name{staticGP}
\alias{staticGP}
\title{Fit a linear model using Gaussian Process.}
\usage{
staticGP(
  datafile = NULL,
  dataref = NULL,
  outcome,
  treatment,
  x.explanatory = NULL,
  x.confounding = NULL,
  tr.hte = NULL,
  tr2.hte = NULL,
  burn.num = 500,
  mcmc.num = 500,
  outcome.lb = NULL,
  outcome.ub = NULL,
  outcome.bound_censor = "neither",
  outcome.type = "Continuous",
  outcome.censor.lv = NULL,
  outcome.censor.uv = NULL,
  outcome.censor.yn = NULL,
  outcome.link = "identity",
  tr.type = "Discrete",
  tr2.type = "Discrete",
  tr.values = NULL,
  tr2.values = NULL,
  pr.values = NULL,
  x.categorical = NULL,
  method = "BART",
  mi.datafile = NULL,
  mi.dataref = NULL,
  sheet = NULL,
  mi.sheet = NULL
)
}
\arguments{
\item{datafile}{File to upload (.csv or .xls)}

\item{dataref}{Reference to already uploaded file.}

\item{outcome}{The name of the outcome variable.}

\item{treatment}{The vector of the name of the treatment variables.}

\item{x.explanatory}{The vector of the name of the explanatory variables.}

\item{x.confounding}{The vector of the name of the confounding variables.}

\item{tr.hte}{An optional vector specifying variables which may have heterogeneous treatment effect with the first treatment variable}

\item{tr2.hte}{An optional vector specifying variables which may have heterogeneous treatment effect with the second treatment variable}

\item{burn.num}{numeric; the number of MCMC 'burn-in' samples, i.e. number of MCMC to be discarded.}

\item{mcmc.num}{numeric; the number of MCMC samples after 'burn-in'.}

\item{outcome.lb}{Lower bound if \code{outcome.censor} is TRUE.}

\item{outcome.ub}{Upper bound if \code{outcome.censor} is TRUE.}

\item{outcome.bound_censor}{logical; if TRUE, outcomes are bounded. If FALSE, outcomes are not bounded.}

\item{outcome.type}{Outcome type ("Continuous" or "Discrete").}

\item{outcome.censor.lv}{outcome.censor.lv}

\item{outcome.censor.uv}{outcome.censor.uv}

\item{outcome.censor.yn}{outcome.censor.yn}

\item{outcome.link}{outcome.link}

\item{tr.type}{tr.type}

\item{tr2.type}{tr2.type}

\item{tr.values}{user-defined values for the calculation of ATE if the first treatment variable is continuous}

\item{tr2.values}{user-defined values for the calculation of ATE if the second treatment variable is continuous}

\item{pr.values}{pr.values}

\item{x.categorical}{x.categorical}

\item{method}{method}

\item{mi.datafile}{File to upload (.csv or .xls) that contains the imputed data in the model.}

\item{mi.dataref}{Reference to already uploaded file that contains the imputed data in the model.}

\item{sheet}{If \code{dataurl} points to Excel file this variable specifies which sheet to load.}

\item{mi.sheet}{If \code{mi.dataurl} points to Excel file this variable specifies which sheet to load.}

\item{outcome.censor}{logical; if TRUE, outcomes are bounded. If FALSE, outcomes are not bounded.}

\item{categorical}{Ensure the specified fields are categorical}
}
\value{
jobid
}
\description{
FIXME: GPMatch is used to fit a linear model within a Bayesian framework.
  Gaussian process (GP) prior covariance function is utilized as a matching tool
  which accomplishes matching and flexible outcome modeling in a single step.
}
